Removing repeated logic for file upload
